AVRO-204. Change the way symbolic references are tracked.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/avro/trunk@881076 13f79535-47bb-0310-9956-ffa450edef68

+++ src/c++/api/Validator.hh
@@ -39,7 +39,7 @@ class OutputStreamer;
 
 class Validator : private boost::noncopyable
 {
-    typedef uint64_t flag_t;
+    typedef uint32_t flag_t;
 
   public:
